requirements.txt
annotated-types
anyio
bidict
blinker
cachetools
certifi
charset-normalizer
click
distro
dnspython
easyocr
eventlet
filelock
Flask
Flask-SocketIO
fsspec
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
googleapis-common-protos
greenlet
groq
gspread
gunicorn
h11
httpcore
httplib2
httpx
idna
imageio
itsdangerous
Jinja2
lazy_loader
lxml
MarkupSafe
mpmath
networkx
ninja
numpy
nvidia-cublas-cu12
nvidia-cuda-cupti-cu12
nvidia-cuda-nvrtc-cu12
nvidia-cuda-runtime-cu12
nvidia-cudnn-cu12
nvidia-cufft-cu12
nvidia-curand-cu12
nvidia-cusolver-cu12
nvidia-cusparse-cu12
nvidia-cusparselt-cu12
nvidia-nccl-cu12
nvidia-nvjitlink-cu12
nvidia-nvtx-cu12
oauth2client
oauthlib
opencv-python-headless
packaging
pandas
pdf2image
pillow
proto-plus
protobuf
pyasn1
pyasn1_modules
pyclipper
pydantic
pydantic_core
PyMuPDF
pyparsing
python-bidi
python-dateutil
python-docx
python-engineio
python-socketio
pytz
PyYAML
requests
requests-oauthlib
rsa
scikit-image
scipy
setuptools
shapely
simple-websocket
six
sniffio
sympy
tifffile
torch
torchvision
triton
typing_extensions
tzdata
uritemplate
urllib3
Werkzeug
wsproto
motor